refactor: reduce the number of modules
